requirements.txt
streamlit>=1.28.0
langchain>=0.0.346
langchain-community>=0.0.10
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pypdf>=3.17.0
python-docx>=1.1.0
tiktoken>=0.5.0
streamlit-chat>=0.1.0